legal-assistant/legal_assistant/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework-simplejwt==5.2.2
google-generativeai==0.3.1
python-dotenv==1.0.0
Pillow==9.5.0